refactor(web): split media upload transfer
