+++ apps/web/src/appData/sync/mediaUploads/mediaUploadTransferRunner.ts
@@ -1,177 +1,56 @@
 import {
-  ApiContractError,
-  ApiError,
-  apiNetworkRetryMaximumAttemptCount,
-  abortMediaAssetUploadSession,
-  completeMediaAssetUploadSession,
-  createApiNetworkRetryDelayMs,
-  createMediaAssetUploadPartUrls,
   createMediaAssetUploadSession,
   isAuthRedirectError,
+  type ApiError,
 } from "../../../api";
 import { putMediaAsset } from "../../../localDb/mediaAssets";
 import {
-  claimNextDueMediaTransferByKind,
   loadMediaBlobCacheRecord,
-  markClaimedMediaTransferFailed,
   markClaimedMediaTransferSucceeded,
-  markMediaUploadTransferCompletionTerminal,
-  recoverStaleInProgressMediaTransfersByKind,
-  renewInProgressMediaTransferClaim,
   type MediaBlobCacheRecord,
   type MediaTransferQueueRecord,
 } from "../../../localDb/mediaTransfers";
 import { loadCloudSettings } from "../../../localDb/sync/cloudSettings";
 import type {
-  CompleteMediaAssetUploadPartInput,
   MediaAsset,
   MediaAssetUploadSessionCreateResult,
-  MediaAssetUploadPartUrl,
   MediaAssetUploadSession,
 } from "../../../types";
 import { requireCloudInstallationId } from "../local/syncCloudSettings";
+import {
+  claimNextDueUploadTransfer,
+  markAuthRedirectUploadTransferRetryable,
+  markUploadTransferCompletionTerminal,
+  markUploadTransferFailed,
+  normalizeMediaUploadError,
+  PermanentMediaUploadError,
+  recoverStaleUploadTransferClaims,
+  startUploadClaimHeartbeat,
+  throwIfUploadLifecycleCancelled,
+  type MediaUploadClaimHeartbeat,
+} from "./mediaUploadClaimLifecycle";
+import {
+  abortUploadSessionAfterFailure,
+  combineUploadFailureWithAbortFailure,
+  completeMultipartUploadSession,
+  isSameSessionCompletionRetryError,
+  MediaUploadCompletionTerminalError,
+  type MediaUploadCompletionResult,
+} from "./multipartCompletion";
+import {
+  uploadParts,
+  type PlannedUploadPart,
+} from "./signedPartUpload";
 
 type VerifiedUploadBytes = Readonly<{
   bytes: Uint8Array;
   blob: Blob;
 }>;
 
-type PlannedUploadPart = Readonly<{
-  partNumber: number;
-  sha256: string;
-  startByte: number;
-  endByte: number;
-}>;
-
-type UploadedMediaPart = Readonly<{
-  partNumber: number;
-  eTag: string;
-  sha256: string;
-}>;
-
-type MediaUploadFailureKind = "retryable" | "permanent";
-
-type MediaUploadFailure = Readonly<{
-  kind: MediaUploadFailureKind;
-  message: string;
-}>;
-
-type MediaUploadClaimHeartbeat = Readonly<{
-  failureSignal: AbortSignal;
-  getClaimedAt: () => string;
-  throwIfFailed: () => Promise<void>;
-  waitForFailure: () => Promise<void>;
-  stop: () => Promise<unknown | null>;
-}>;
-
-type MediaUploadCompletionResult = Readonly<{
-  mediaAsset: MediaAsset;
-  retryableCompletionCause: ApiError | null;
-}>;
-
-class RetryableMediaUploadError extends Error {
-  constructor(message: string) {
-    super(message);
-    this.name = "RetryableMediaUploadError";
-  }
-}
-
-class PermanentMediaUploadError extends Error {
-  constructor(message: string) {
-    super(message);
-    this.name = "PermanentMediaUploadError";
-  }
-}
-
-type MediaUploadCompletionTerminalReason = "retry_exhausted" | "interrupted";
-
-class MediaUploadCompletionTerminalError extends PermanentMediaUploadError {
-  readonly reason: MediaUploadCompletionTerminalReason;
-  readonly completionCause: ApiError;
-  readonly interruptionCause: Error | null;
-
-  constructor(
-    reason: MediaUploadCompletionTerminalReason,
-    completionCause: ApiError,
-    interruptionCause: Error | null,
-  ) {
-    const interruptionDescription = interruptionCause === null
-      ? "none"
-      : describeUploadError(interruptionCause);
-    super(
-      `Media upload completion stopped for this local run: reason=${reason}, `
-      + `completionError=${describeApiError(completionCause)}, `
-      + `interruptionError=${interruptionDescription}`,
-    );
-    this.name = "MediaUploadCompletionTerminalError";
-    this.reason = reason;
-    this.completionCause = completionCause;
-    this.interruptionCause = interruptionCause;
-  }
-}
-
 const browserMediaUploadPartSizeBytes = 8 * 1024 * 1024;
-const retryBaseDelayMs = 30_000;
-const retryMaximumDelayMs = 30 * 60_000;
-const signedPartUploadMaximumAttemptCount = 2;
-const signedPartUploadRetryDelayMs = 250;
-const signedPartUrlExpirySafetyMarginMs = 60_000;
-const signedPartUploadErrorBodyMaximumBytes = 4096;
-const signedPartUploadErrorFieldMaximumLength = 128;
-const staleMediaUploadClaimLeaseMs = 30 * 60_000;
-const mediaUploadClaimHeartbeatMs = 5 * 60_000;
-const permanentlyFailedNextAttemptAt = "9999-12-31T23:59:59.999Z";
-const retryableUploadSessionErrorCodes: ReadonlySet<string> = new Set([
-  "MEDIA_ASSET_STORAGE_UNAVAILABLE",
-  "MEDIA_ASSET_UPLOAD_NOT_FOUND",
-  "MEDIA_ASSET_UPLOAD_SESSION_EXPIRED",
-  "MEDIA_ASSET_UPLOAD_SESSION_NOT_FOUND",
-  "MEDIA_ASSET_UPLOAD_SESSION_RECOVERY_FAILED",
-]);
-const sameSessionCompletionRetryErrorCodes: ReadonlySet<string> = new Set([
-  "MEDIA_ASSET_UPLOAD_SESSION_COMPLETION_DEADLINE_EXCEEDED",
-  "MEDIA_ASSET_UPLOAD_SESSION_COMPLETION_IN_PROGRESS",
-]);
 
 function isBrowserOnline(): boolean {
   return typeof navigator === "undefined" || navigator.onLine !== false;
-}
-
-function readErrorName(error: unknown): string {
-  if (error instanceof Error && error.name.trim() !== "") {
-    return error.name;
-  }
-
-  return typeof error;
-}
-
-function readUploadLifecycleCancellationError(signal: AbortSignal): Error {
-  const reason: unknown = signal.reason;
-  if (reason instanceof Error) {
-    return reason;
-  }
-  if (typeof reason === "string" && reason.trim() !== "") {
-    return new Error(reason);
-  }
-  return new Error("Media upload lifecycle was cancelled");
-}
-
-function throwIfUploadLifecycleCancelled(signal: AbortSignal): void {
-  if (signal.aborted) {
-    throw readUploadLifecycleCancellationError(signal);
-  }
-}
-
-function readErrorMessage(error: unknown): string {
-  if (error instanceof Error && error.message.trim() !== "") {
-    return error.message;
-  }
-
-  return String(error);
-}
-
-function normalizeMediaUploadError(error: unknown): Error {
-  return error instanceof Error ? error : new Error(readErrorMessage(error));
 }
 
 function bytesToHex(bytes: Uint8Array): string {
@@ -362,746 +241,6 @@ async function buildPlannedUploadParts(
   return uploadParts;
 }
 
-function isTransientStatusCode(statusCode: number): boolean {
-  return statusCode === 408 || statusCode === 429 || statusCode >= 500;
-}
-
-function isRetryableApiError(error: ApiError): boolean {
-  if (error.statusCode === 0 || isTransientStatusCode(error.statusCode)) {
-    return true;
-  }
-
-  return error.code !== null && retryableUploadSessionErrorCodes.has(error.code);
-}
-
-function describeApiError(error: ApiError): string {
-  return [
-    error.message,
-    `endpoint=${error.endpoint}`,
-    `status=${error.statusCode}`,
-    `code=${error.code ?? "none"}`,
-    `requestId=${error.requestId ?? "none"}`,
-    `responseBodyKind=${error.responseBodyKind}`,
-  ].join(" ");
-}
-
-function describeUploadError(error: unknown): string {
-  if (error instanceof ApiError) {
-    return describeApiError(error);
-  }
-
-  return readErrorMessage(error);
-}
-
-function classifyMediaUploadError(error: unknown): MediaUploadFailure {
-  if (error instanceof RetryableMediaUploadError) {
-    return {
-      kind: "retryable",
-      message: error.message,
-    };
-  }
-
-  if (error instanceof PermanentMediaUploadError || error instanceof ApiContractError) {
-    return {
-      kind: "permanent",
-      message: describeUploadError(error),
-    };
-  }
-
-  if (error instanceof ApiError) {
-    return {
-      kind: isRetryableApiError(error) ? "retryable" : "permanent",
-      message: describeApiError(error),
-    };
-  }
-
-  return {
-    kind: "retryable",
-    message: readErrorMessage(error),
-  };
-}
-
-function describeUploadSessionCleanupFailure(primaryFailure: MediaUploadFailure, abortFailure: MediaUploadFailure): string {
-  return `Media upload session cleanup failed after upload error: primaryError=${primaryFailure.message}, abortError=${abortFailure.message}`;
-}
-
-function combineUploadFailureWithAbortFailure(primaryError: unknown, abortError: unknown | null): unknown {
-  if (abortError === null || isAuthRedirectError(primaryError)) {
-    return primaryError;
-  }
-
-  if (isAuthRedirectError(abortError)) {
-    return abortError;
-  }
-
-  const primaryFailure = classifyMediaUploadError(primaryError);
-  const abortFailure = classifyMediaUploadError(abortError);
-  if (primaryFailure.kind === "retryable" || abortFailure.kind === "retryable") {
-    return new RetryableMediaUploadError(describeUploadSessionCleanupFailure(primaryFailure, abortFailure));
-  }
-
-  return new PermanentMediaUploadError(describeUploadSessionCleanupFailure(primaryFailure, abortFailure));
-}
-
-function createRetryNextAttemptAt(transfer: MediaTransferQueueRecord, failedAt: string): string {
-  const nextAttemptCount = transfer.attemptCount + 1;
-  const uncappedDelayMs = retryBaseDelayMs * (2 ** Math.max(0, nextAttemptCount - 1));
-  const delayMs = Math.min(uncappedDelayMs, retryMaximumDelayMs);
-  return new Date(new Date(failedAt).getTime() + delayMs).toISOString();
-}
-
-function createStaleUploadClaimCutoff(recoveredAt: string): string {
-  return new Date(new Date(recoveredAt).getTime() - staleMediaUploadClaimLeaseMs).toISOString();
-}
-
-function requireInProgressUploadClaimedAt(transfer: MediaTransferQueueRecord): string {
-  if (transfer.claimedAt === null) {
-    throw new PermanentMediaUploadError(`Media upload transfer claim is missing: transferId=${transfer.transferId}`);
-  }
-
-  return transfer.claimedAt;
-}
-
-async function renewUploadTransferClaim(
-  transfer: MediaTransferQueueRecord,
-  expectedClaimedAt: string,
-): Promise<string> {
-  const renewedAt = new Date().toISOString();
-  const renewedTransfer = await renewInProgressMediaTransferClaim({
-    transferId: transfer.transferId,
-    kind: "upload",
-    expectedClaimedAt,
-    renewedAt,
-  });
-  return requireInProgressUploadClaimedAt(renewedTransfer);
-}
-
-function startUploadClaimHeartbeat(transfer: MediaTransferQueueRecord): MediaUploadClaimHeartbeat {
-  let claimedAt = requireInProgressUploadClaimedAt(transfer);
-  let heartbeatError: unknown = null;
-  let renewalTask: Promise<void> = Promise.resolve();
-  let didStop = false;
-  let resolveHeartbeatFailure: (() => void) | null = null;
-  const heartbeatFailureController = new AbortController();
-  const heartbeatFailurePromise = new Promise<void>((resolve) => {
-    resolveHeartbeatFailure = resolve;
-  });
-
-  const queueRenewal = (): void => {
-    if (heartbeatError !== null || didStop) {
-      return;
-    }
-
-    renewalTask = renewalTask
-      .then(async (): Promise<void> => {
-        claimedAt = await renewUploadTransferClaim(transfer, claimedAt);
-      })
-      .catch((error: unknown): void => {
-        heartbeatError = error;
-        heartbeatFailureController.abort(error);
-        resolveHeartbeatFailure?.();
-      });
-  };
-
-  const timerId = window.setInterval(queueRenewal, mediaUploadClaimHeartbeatMs);
-  queueRenewal();
-
-  return {
-    failureSignal: heartbeatFailureController.signal,
-    getClaimedAt: () => claimedAt,
-    throwIfFailed: async (): Promise<void> => {
-      await renewalTask;
-      if (heartbeatError !== null) {
-        throw heartbeatError;
-      }
-    },
-    waitForFailure: () => heartbeatFailurePromise,
-    stop: async (): Promise<unknown | null> => {
-      if (didStop === false) {
-        didStop = true;
-        window.clearInterval(timerId);
-      }
-
-      await renewalTask;
-      return heartbeatError;
-    },
-  };
-}
-
-async function recoverStaleUploadTransferClaims(workspaceId: string, recoveredAt: string): Promise<void> {
-  const staleClaimedBefore = createStaleUploadClaimCutoff(recoveredAt);
-  await recoverStaleInProgressMediaTransfersByKind({
-    workspaceId,
-    kind: "upload",
-    staleClaimedBefore,
-    recoveredAt,
-    nextAttemptAt: recoveredAt,
-    lastError: `Media upload transfer reclaimed after stale in-progress claim: staleClaimedBefore=${staleClaimedBefore}, recoveredAt=${recoveredAt}`,
-  });
-}
-
-async function markUploadTransferFailed(
-  transfer: MediaTransferQueueRecord,
-  claimedAt: string,
-  error: unknown,
-): Promise<void> {
-  const failedAt = new Date().toISOString();
-  const failure = classifyMediaUploadError(error);
-  const nextAttemptAt = failure.kind === "retryable"
-    ? createRetryNextAttemptAt(transfer, failedAt)
-    : permanentlyFailedNextAttemptAt;
-  const lastError = `Media upload transfer failed (${failure.kind}): transferId=${transfer.transferId}, workspaceId=${transfer.workspaceId}, mediaAssetId=${transfer.mediaAssetId}, error=${failure.message}`;
-  if (error instanceof MediaUploadCompletionTerminalError) {
-    await markMediaUploadTransferCompletionTerminal({
-      transferId: transfer.transferId,
-      workspaceId: transfer.workspaceId,
-      mediaAssetId: transfer.mediaAssetId,
-      failedAt,
-      lastError,
-      nextAttemptAt: permanentlyFailedNextAttemptAt,
-    });
-    return;
-  }
-
-  await markClaimedMediaTransferFailed({
-    transferId: transfer.transferId,
-    kind: "upload",
-    expectedClaimedAt: claimedAt,
-    failedAt,
-    lastError,
-    nextAttemptAt,
-  });
-}
-
-async function markAuthRedirectUploadTransferRetryable(
-  transfer: MediaTransferQueueRecord,
-  claimedAt: string,
-): Promise<void> {
-  const failedAt = new Date().toISOString();
-  await markClaimedMediaTransferFailed({
-    transferId: transfer.transferId,
-    kind: "upload",
-    expectedClaimedAt: claimedAt,
-    failedAt,
-    lastError: `Media upload transfer paused for browser authentication: transferId=${transfer.transferId}, workspaceId=${transfer.workspaceId}, mediaAssetId=${transfer.mediaAssetId}`,
-    nextAttemptAt: failedAt,
-  });
-}
-
-async function waitForSignedPartUploadRetry(signal: AbortSignal): Promise<void> {
-  await new Promise<void>((resolve, reject) => {
-    let timerId: number | null = null;
-    const abortHandler = (): void => {
-      if (timerId !== null) {
-        window.clearTimeout(timerId);
-      }
-      reject(readUploadLifecycleCancellationError(signal));
-    };
-    if (signal.aborted) {
-      abortHandler();
-      return;
-    }
-
-    timerId = window.setTimeout((): void => {
-      signal.removeEventListener("abort", abortHandler);
-      resolve();
-    }, signedPartUploadRetryDelayMs);
-    signal.addEventListener("abort", abortHandler, { once: true });
-  });
-}
-
-function warnSignedPartUploadRetry(
-  transfer: MediaTransferQueueRecord,
-  partNumber: number,
-  attemptNumber: number,
-  error: unknown,
-): void {
-  console.warn("Media upload part retry", {
-    transferId: transfer.transferId,
-    workspaceId: transfer.workspaceId,
-    mediaAssetId: transfer.mediaAssetId,
-    partNumber,
-    attemptNumber,
-    nextAttemptNumber: attemptNumber + 1,
-    errorName: readErrorName(error),
-    errorMessage: readErrorMessage(error),
-  });
-}
-
-async function readSignedPartUploadFailureBodyText(response: Response): Promise<string | null> {
-  const responseBody = response.body;
-  if (responseBody === null) {
-    return null;
-  }
-
-  const reader = responseBody.getReader();
-  const chunks: Array<Uint8Array> = [];
-  let totalByteLength = 0;
-  try {
-    while (totalByteLength < signedPartUploadErrorBodyMaximumBytes) {
-      const result = await reader.read();
-      if (result.done) {
-        break;
-      }
-
-      const remainingByteLength = signedPartUploadErrorBodyMaximumBytes - totalByteLength;
-      const chunk = result.value.byteLength > remainingByteLength
-        ? result.value.slice(0, remainingByteLength)
-        : result.value;
-      chunks.push(chunk);
-      totalByteLength += chunk.byteLength;
-      if (result.value.byteLength >= remainingByteLength) {
-        await reader.cancel();
-        break;
-      }
-    }
-  } catch (error) {
-    console.warn("Media upload part failure body read failed", {
-      errorName: readErrorName(error),
-      errorMessage: toBoundedSignedPartErrorField(readErrorMessage(error)) ?? "none",
-    });
-    return null;
-  } finally {
-    try {
-      reader.releaseLock();
-    } catch {
-      // The stream may already be released after cancellation.
-    }
-  }
-
-  const bodyBytes = new Uint8Array(totalByteLength);
-  let offset = 0;
-  for (const chunk of chunks) {
-    bodyBytes.set(chunk, offset);
-    offset += chunk.byteLength;
-  }
-
-  return new TextDecoder().decode(bodyBytes);
-}
-
-type SanitizedSignedPartUploadFailure = Readonly<{
-  status: number;
-  statusText: string;
-  s3Code: string | null;
-  requestId: string | null;
-  extendedRequestId: string | null;
-}>;
-
-function readXmlElementText(documentValue: Document, tagName: string): string | null {
-  const element = documentValue.getElementsByTagName(tagName)[0];
-  if (element === undefined) {
-    return null;
-  }
-
-  const text = element.textContent?.trim() ?? "";
-  return text === "" ? null : text;
-}
-
-function toBoundedSignedPartErrorField(value: string | null): string | null {
-  if (value === null) {
-    return null;
-  }
-
-  const trimmedValue = value.trim();
-  if (trimmedValue === "") {
-    return null;
-  }
-
-  if (trimmedValue.length <= signedPartUploadErrorFieldMaximumLength) {
-    return trimmedValue;
-  }
-
-  return `${trimmedValue.slice(0, signedPartUploadErrorFieldMaximumLength)}...`;
-}
-
-function parseS3ErrorBody(responseBody: string | null): Pick<SanitizedSignedPartUploadFailure, "s3Code" | "requestId" | "extendedRequestId"> {
-  if (responseBody === null || responseBody.trim() === "" || typeof DOMParser === "undefined") {
-    return {
-      s3Code: null,
-      requestId: null,
-      extendedRequestId: null,
-    };
-  }
-
-  const documentValue = new DOMParser().parseFromString(responseBody, "application/xml");
-  if (documentValue.getElementsByTagName("parsererror").length > 0) {
-    return {
-      s3Code: null,
-      requestId: null,
-      extendedRequestId: null,
-    };
-  }
-
-  return {
-    s3Code: readXmlElementText(documentValue, "Code"),
-    requestId: readXmlElementText(documentValue, "RequestId"),
-    extendedRequestId: readXmlElementText(documentValue, "HostId"),
-  };
-}
-
-async function readSanitizedSignedPartUploadFailure(response: Response): Promise<SanitizedSignedPartUploadFailure> {
-  const responseBody = await readSignedPartUploadFailureBodyText(response);
-  const parsedBody = parseS3ErrorBody(responseBody);
-  return {
-    status: response.status,
-    statusText: toBoundedSignedPartErrorField(response.statusText) ?? "",
-    s3Code: toBoundedSignedPartErrorField(parsedBody.s3Code),
-    requestId: toBoundedSignedPartErrorField(response.headers.get("x-amz-request-id") ?? parsedBody.requestId),
-    extendedRequestId: toBoundedSignedPartErrorField(response.headers.get("x-amz-id-2") ?? parsedBody.extendedRequestId),
-  };
-}
-
-function formatNullableSignedPartErrorField(value: string | null): string {
-  return value ?? "none";
-}
-
-function describeSanitizedSignedPartUploadFailure(
-  transfer: MediaTransferQueueRecord,
-  part: PlannedUploadPart,
-  failure: SanitizedSignedPartUploadFailure,
-): string {
-  return [
-    "Media upload part failed:",
-    `transferId=${transfer.transferId}`,
-    `partNumber=${part.partNumber}`,
-    `status=${failure.status}`,
-    `statusText=${failure.statusText}`,
-    `s3Code=${formatNullableSignedPartErrorField(failure.s3Code)}`,
-    `requestId=${formatNullableSignedPartErrorField(failure.requestId)}`,
-    `extendedRequestId=${formatNullableSignedPartErrorField(failure.extendedRequestId)}`,
-  ].join(" ");
-}
-
-function assertSignedPartUrlUsable(
-  transfer: MediaTransferQueueRecord,
-  partUrl: MediaAssetUploadPartUrl,
-  part: PlannedUploadPart,
-): void {
-  const expiresAtTime = Date.parse(partUrl.expiresAt);
-  if (Number.isFinite(expiresAtTime) === false) {
-    throw new PermanentMediaUploadError(`Media upload part URL expiration is invalid: transferId=${transfer.transferId}, partNumber=${part.partNumber}, expiresAt=${partUrl.expiresAt}`);
-  }
-
-  if (expiresAtTime - Date.now() <= signedPartUrlExpirySafetyMarginMs) {
-    throw new RetryableMediaUploadError(`Media upload part URL expires too soon: transferId=${transfer.transferId}, partNumber=${part.partNumber}, expiresAt=${partUrl.expiresAt}, safetyMarginMs=${signedPartUrlExpirySafetyMarginMs}`);
-  }
-}
-
-function warnUploadSessionAbortFailure(
-  transfer: MediaTransferQueueRecord,
-  sessionId: string,
-  error: unknown,
-): void {
-  console.warn("Media upload session abort failed", {
-    transferId: transfer.transferId,
-    workspaceId: transfer.workspaceId,
-    mediaAssetId: transfer.mediaAssetId,
-    sessionId,
-    errorName: readErrorName(error),
-    errorMessage: describeUploadError(error),
-  });
-}
-
-async function uploadSignedPartOnce(
-  transfer: MediaTransferQueueRecord,
-  partUrl: MediaAssetUploadPartUrl,
-  part: PlannedUploadPart,
-  blob: Blob,
-  signal: AbortSignal,
-): Promise<string> {
-  assertSignedPartUrlUsable(transfer, partUrl, part);
-  let response: Response;
-  try {
-    response = await fetch(partUrl.url, {
-      method: partUrl.method,
-      headers: partUrl.headers,
-      body: blob.slice(part.startByte, part.endByte),
-      signal,
-    });
-  } catch (error) {
-    throw new RetryableMediaUploadError(`Media upload part request failed: transferId=${transfer.transferId}, partNumber=${part.partNumber}, errorName=${readErrorName(error)}`);
-  }
-
-  if (response.ok === false) {
-    const failure = await readSanitizedSignedPartUploadFailure(response);
-    const message = describeSanitizedSignedPartUploadFailure(transfer, part, failure);
-    if (isTransientStatusCode(response.status) || response.status === 403) {
-      throw new RetryableMediaUploadError(message);
-    }
-
-    throw new PermanentMediaUploadError(message);
-  }
-
-  const eTag = response.headers.get("ETag");
-  if (eTag === null || eTag.trim() === "") {
-    throw new PermanentMediaUploadError(`Media upload part response missing ETag: transferId=${transfer.transferId}, partNumber=${part.partNumber}, status=${response.status}`);
-  }
-
-  return eTag;
-}
-
-async function uploadSignedPart(
-  transfer: MediaTransferQueueRecord,
-  sessionId: string,
-  part: PlannedUploadPart,
-  blob: Blob,
-  heartbeat: MediaUploadClaimHeartbeat,
-  signal: AbortSignal,
-): Promise<UploadedMediaPart> {
-  let attemptNumber = 1;
-  while (true) {
-    try {
-      const partUrl = await loadPartUrl(transfer, sessionId, part, signal);
-      await heartbeat.throwIfFailed();
-      return {
-        partNumber: part.partNumber,
-        eTag: await uploadSignedPartOnce(transfer, partUrl, part, blob, signal),
-        sha256: part.sha256,
-      };
-    } catch (error) {
-      if (error instanceof RetryableMediaUploadError === false || attemptNumber >= signedPartUploadMaximumAttemptCount) {
-        throw error;
-      }
-
-      warnSignedPartUploadRetry(transfer, part.partNumber, attemptNumber, error);
-      await waitForSignedPartUploadRetry(signal);
-      attemptNumber += 1;
-    }
-  }
-}
-
-function getPartUrl(
-  partUrls: ReadonlyArray<MediaAssetUploadPartUrl>,
-  part: PlannedUploadPart,
-  transfer: MediaTransferQueueRecord,
-): MediaAssetUploadPartUrl {
-  const matchingPartUrls = partUrls.filter((partUrl) => partUrl.partNumber === part.partNumber);
-  if (matchingPartUrls.length !== 1) {
-    throw new PermanentMediaUploadError(`Media upload part URL response mismatch: transferId=${transfer.transferId}, partNumber=${part.partNumber}, matchingUrlCount=${matchingPartUrls.length}`);
-  }
-
-  return matchingPartUrls[0];
-}
-
-async function loadPartUrls(
-  transfer: MediaTransferQueueRecord,
-  sessionId: string,
-  parts: ReadonlyArray<PlannedUploadPart>,
-  signal: AbortSignal,
-): Promise<ReadonlyArray<MediaAssetUploadPartUrl>> {
-  const response = await createMediaAssetUploadPartUrls(transfer.workspaceId, sessionId, {
-    parts: parts.map((part) => ({
-      partNumber: part.partNumber,
-      sha256: part.sha256,
-    })),
-  }, signal);
-  if (response.sessionId !== sessionId) {
-    throw new PermanentMediaUploadError(`Media upload part URL session mismatch: transferId=${transfer.transferId}, expectedSessionId=${sessionId}, actualSessionId=${response.sessionId}`);
-  }
-
-  return parts.map((part) => getPartUrl(response.partUrls, part, transfer));
-}
-
-async function loadPartUrl(
-  transfer: MediaTransferQueueRecord,
-  sessionId: string,
-  part: PlannedUploadPart,
-  signal: AbortSignal,
-): Promise<MediaAssetUploadPartUrl> {
-  const [partUrl] = await loadPartUrls(transfer, sessionId, [part], signal);
-  if (partUrl === undefined) {
-    throw new PermanentMediaUploadError(`Media upload part URL response was empty: transferId=${transfer.transferId}, partNumber=${part.partNumber}`);
-  }
-
-  return partUrl;
-}
-
-async function uploadParts(
-  transfer: MediaTransferQueueRecord,
-  uploadSession: MediaAssetUploadSession,
-  parts: ReadonlyArray<PlannedUploadPart>,
-  blob: Blob,
-  heartbeat: MediaUploadClaimHeartbeat,
-  signal: AbortSignal,
-): Promise<ReadonlyArray<UploadedMediaPart>> {
-  const uploadedParts: Array<UploadedMediaPart> = [];
-  for (const part of parts) {
-    await heartbeat.throwIfFailed();
-    uploadedParts.push(await uploadSignedPart(
-      transfer,
-      uploadSession.sessionId,
-      part,
-      blob,
-      heartbeat,
-      signal,
-    ));
-  }
-
-  return uploadedParts;
-}
-
-function toCompleteParts(uploadedParts: ReadonlyArray<UploadedMediaPart>): ReadonlyArray<CompleteMediaAssetUploadPartInput> {
-  return uploadedParts.map((part) => ({
-    partNumber: part.partNumber,
-    eTag: part.eTag,
-    sha256: part.sha256,
-  }));
-}
-
-function isSameSessionCompletionRetryError(error: unknown): error is ApiError {
-  return error instanceof ApiError
-    && error.code !== null
-    && sameSessionCompletionRetryErrorCodes.has(error.code);
-}
-
-async function waitForCompletionRetry(
-  delayMs: number,
-  heartbeat: MediaUploadClaimHeartbeat,
-  signal: AbortSignal,
-): Promise<void> {
-  let timerId: number | null = null;
-  let abortHandler: (() => void) | null = null;
-  try {
-    await Promise.race([
-      new Promise<void>((resolve) => {
-        timerId = window.setTimeout(resolve, delayMs);
-      }),
-      heartbeat.waitForFailure(),
-      new Promise<void>((_resolve, reject) => {
-        abortHandler = () => {
-          reject(readUploadLifecycleCancellationError(signal));
-        };
-        if (signal.aborted) {
-          abortHandler();
-          return;
-        }
-        signal.addEventListener("abort", abortHandler, { once: true });
-      }),
-    ]);
-  } finally {
-    if (timerId !== null) {
-      window.clearTimeout(timerId);
-    }
-    if (abortHandler !== null) {
-      signal.removeEventListener("abort", abortHandler);
-    }
-  }
-
-  throwIfUploadLifecycleCancelled(signal);
-  await heartbeat.throwIfFailed();
-  throwIfUploadLifecycleCancelled(signal);
-}
-
-async function validateCompletionOwnership(
-  heartbeat: MediaUploadClaimHeartbeat,
-  signal: AbortSignal,
-  retryableCompletionCause: ApiError | null,
-): Promise<void> {
-  try {
-    throwIfUploadLifecycleCancelled(signal);
-    await heartbeat.throwIfFailed();
-    throwIfUploadLifecycleCancelled(signal);
-  } catch (error) {
-    if (retryableCompletionCause !== null) {
-      throw new MediaUploadCompletionTerminalError(
-        "interrupted",
-        retryableCompletionCause,
-        normalizeMediaUploadError(error),
-      );
-    }
-    throw error;
-  }
-}
-
-async function completeMultipartUploadSession(
-  transfer: MediaTransferQueueRecord,
-  uploadSession: MediaAssetUploadSession,
-  uploadedParts: ReadonlyArray<UploadedMediaPart>,
-  heartbeat: MediaUploadClaimHeartbeat,
-  signal: AbortSignal,
-): Promise<MediaUploadCompletionResult> {
-  const request = {
-    parts: toCompleteParts(uploadedParts),
-  };
-  const completionSignal = AbortSignal.any([signal, heartbeat.failureSignal]);
-  let attemptNumber = 1;
-  let lastRetryableCompletionError: ApiError | null = null;
-
-  while (true) {
-    await validateCompletionOwnership(heartbeat, signal, lastRetryableCompletionError);
-
-    try {
-      const result = await completeMediaAssetUploadSession(
-        transfer.workspaceId,
-        uploadSession.sessionId,
-        request,
-        completionSignal,
-      );
-      await validateCompletionOwnership(heartbeat, signal, lastRetryableCompletionError);
-      return {
-        mediaAsset: result.mediaAsset,
-        retryableCompletionCause: lastRetryableCompletionError,
-      };
-    } catch (error) {
-      if (error instanceof MediaUploadCompletionTerminalError) {
-        throw error;
-      }
-      if (isSameSessionCompletionRetryError(error) === false) {
-        await validateCompletionOwnership(heartbeat, signal, lastRetryableCompletionError);
-        if (lastRetryableCompletionError !== null) {
-          throw new MediaUploadCompletionTerminalError(
-            "interrupted",
-            lastRetryableCompletionError,
-            normalizeMediaUploadError(error),
-          );
-        }
-        throw error;
-      }
-      lastRetryableCompletionError = error;
-      await validateCompletionOwnership(heartbeat, signal, lastRetryableCompletionError);
-      if (attemptNumber >= apiNetworkRetryMaximumAttemptCount) {
-        throw new MediaUploadCompletionTerminalError("retry_exhausted", error, null);
-      }
-
-      const delayMs = error.retryAfterMs ?? createApiNetworkRetryDelayMs(attemptNumber);
-      console.warn("Media upload completion retry", {
-        transferId: transfer.transferId,
-        workspaceId: transfer.workspaceId,
-        mediaAssetId: transfer.mediaAssetId,
-        sessionId: uploadSession.sessionId,
-        code: error.code,
-        attemptNumber,
-        maximumAttemptCount: apiNetworkRetryMaximumAttemptCount,
-        nextAttemptNumber: attemptNumber + 1,
-        delayMs,
-      });
-      try {
-        await waitForCompletionRetry(delayMs, heartbeat, signal);
-      } catch (interruptionError) {
-        throw new MediaUploadCompletionTerminalError(
-          "interrupted",
-          error,
-          normalizeMediaUploadError(interruptionError),
-        );
-      }
-      attemptNumber += 1;
-    }
-  }
-}
-
-async function abortUploadSessionAfterFailure(
-  transfer: MediaTransferQueueRecord,
-  sessionId: string,
-): Promise<unknown | null> {
-  try {
-    await abortMediaAssetUploadSession(transfer.workspaceId, sessionId);
-    return null;
-  } catch (error) {
-    warnUploadSessionAbortFailure(transfer, sessionId, error);
-    return error;
-  }
-}
-
 async function runMultipartUploadSession(
   transfer: MediaTransferQueueRecord,
   uploadSession: MediaAssetUploadSession,
@@ -1231,6 +370,10 @@ async function processClaimedUploadTransfer(
       throw failureError;
     }
 
+    if (failureError instanceof MediaUploadCompletionTerminalError) {
+      await markUploadTransferCompletionTerminal(transfer, failureError);
+      return;
+    }
     await markUploadTransferFailed(transfer, heartbeat.getClaimedAt(), failureError);
   }
 }
@@ -1258,7 +401,7 @@ export async function processDueMediaUploadTransfersForWorkspace(
     if (signal.aborted) {
       return;
     }
-    const transfer = await claimNextDueMediaTransferByKind(workspaceId, "upload", new Date().toISOString());
+    const transfer = await claimNextDueUploadTransfer(workspaceId, new Date().toISOString());
     if (transfer === null) {
       return;
     }
